Data request improvement Fetch data at build time
